% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meals_by_category.R
\name{get_meals_by_category}
\alias{get_meals_by_category}
\title{Get Meals by Category}
\usage{
get_meals_by_category(category)
}
\arguments{
\item{category}{A character string specifying the category.}
}
\value{
A dataframe containing meal names and their IDs.
}
\description{
Retrieves meals belonging to a specific category.
}
\examples{
get_meals_by_category("Seafood")
}
