% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meals_by_region.R
\name{get_meals_by_region}
\alias{get_meals_by_region}
\title{Get Meals by Region}
\usage{
get_meals_by_region(region)
}
\arguments{
\item{region}{A character string specifying the region.}
}
\value{
A dataframe containing meal names and their IDs.
}
\description{
Retrieves meals from a specific region/country.
}
\examples{
get_meals_by_region("Italian")
}
