% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_meal.R
\name{get_random_meal}
\alias{get_random_meal}
\title{Get a Random Meal}
\usage{
get_random_meal()
}
\value{
A dataframe containing details of a randomly selected meal.
}
\description{
This function fetches a random meal from TheMealDB API.
}
\examples{
get_random_meal()
}
