% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meal_by_id.R
\name{get_meal_by_id}
\alias{get_meal_by_id}
\title{Get Meal Details by ID}
\usage{
get_meal_by_id(id)
}
\arguments{
\item{id}{A character string representing the meal ID.}
}
\value{
A dataframe containing meal details.
}
\description{
Fetches detailed meal information using a unique ID.
}
\examples{
get_meal_by_id("52772")
}
