% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_meal.R
\name{get_random_meal}
\alias{get_random_meal}
\title{Get a Random Meal}
\usage{
get_random_meal()
}
\value{
A list containing meal data as returned by the API, or NULL if no meal is returned.
}
\description{
Fetches a random meal from TheMealDB API.
}
\examples{
get_random_meal()
}
