% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_meal.R
\name{search_meal}
\alias{search_meal}
\title{Search for a Meal by Name}
\usage{
search_meal(query)
}
\arguments{
\item{query}{A character string with the meal name to search for.}
}
\value{
A dataframe containing meals that match the search query.
}
\description{
This function searches for meals in TheMealDB API by a given name.
}
\examples{
search_meal("pasta")

}
